# gen-go/models/get_commit_information.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// GetCommitInformation get commit information
//
// swagger:model GetCommitInformation
type GetCommitInformation struct {

	// commit sha
	// Required: true
	CommitSha *string `json:"commit_sha"`

	// repo name
	// Required: true
	RepoName *string `json:"repo_name"`
}

// Validate validates this get commit information
func (m *GetCommitInformation) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateCommitSha(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateRepoName(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *GetCommitInformation) validateCommitSha(formats strfmt.Registry) error {

	if err := validate.Required("commit_sha", "body", m.CommitSha); err != nil {
		return err
	}

	return nil
}

func (m *GetCommitInformation) validateRepoName(formats strfmt.Registry) error {

	if err := validate.Required("repo_name", "body", m.RepoName); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *GetCommitInformation) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *GetCommitInformation) UnmarshalBinary(b []byte) error {
	var res GetCommitInformation
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}